made a new file for the controller-test
